We add the default avatar we gonna use.
In utils, we add the historic sorting and the comments sorting. We also add to the Vue instance $saver, where we can save files with javascript. Can be blobs, or URL where there are blobs generated

We add 2 classes to the main css

In TicketView.vue, we prepare the view of an issue with the details, the status, the HistoricTimeline.vue, PeopleRelatedTicket.vue and the comments section. Just missing edit issue info like description and name.

We add also the default icons of files uploaded in statics