In the CategoryCreator.vue, we fix to not modify the prop show, instead we use the emit, to the parent be able to know when the category has been created
PrivateLayout.vue we add the menu of your account in the top right, (not functional yet !)

We add the ControlPanel.vue of the admin, where can see all the users in the app, gonna be able to see all the categories and the statuses. Also gonna be able to edit them and delete if there is not connections with open issues.

We add 2 vies with nothing inside yet as ControlPanel.vue for workers and TodoView for workers (technicians)

In the Login.vue, we add that if its the firstime you login, to set the root password.